backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
websockets>=12
chess>=1.11
pydantic>=2.7
pydantic-settings>=2.3
pyyaml>=6.0
httpx>=0.27
stockfish>=3.28
python-multipart>=0.0.9
